Revert "Set the namelist as the absolute arbiter of whether we are running FABM or not."
We can't compile OFFLINE_FABM and set FABM_MODEL to F in the namelist
and expect FVCOM to advect stuff as OFFLINE_FABM ifdefs out all the
advection routines. So, leaving this check in makes sense.

I am, however, tempted to add "YOU IDIOT" to the end of the error
message...

This reverts commit 53d8db38d62385d9fa3d1e74a66174df7b94a5b2.