Tweaked floating-point comparison in mod_input.F to allow smaller triangles.
This is useful when using a lat/lon mesh with small (~50m) triangles
Otherwise FVCOM tries to check that the nodes are specified clockwise,
and concluse that the triangles are actually lines.

There's a chance that this might break this test for a malformed mesh
if running with single precision. I'm not expert enough to know!